+++ Assets/TowerGuardian/Scripts/Quests/EnterFourthLevelQuest.cs
@@ -30,7 +30,7 @@ namespace TowerGuardian.Scripts.Quests
 
             foreach (var levelData in YG2.saves.LevelsProgress)
             {
-                if (levelData.Level == LevelID.Level4 && levelData.IsComplite)
+                if (levelData.Level == (int) LevelID.Level4 && levelData.IsComplite)
                 {
                     Complete();
                     break;
