+++ Assets/TowerGuardian/Scripts/Quests/EnterThirdLevelQuest.cs
@@ -30,7 +30,7 @@ namespace TowerGuardian.Scripts.Quests
 
             foreach (var levelData in YG2.saves.LevelsProgress)
             {
-                if (levelData.Level == LevelID.Level3 && levelData.IsComplite)
+                if (levelData.Level == (int) LevelID.Level3 && levelData.IsComplite)
                 {
                     Complete();
                     break;
